+++ src/components/NetworkSwitcher/NetworkSwitcher.tsx
@@ -1,12 +1,14 @@
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { networks } from 'config';
 import { networkSelector } from 'redux/selectors';
-import { changeNetwork } from 'redux/slices';
+import { useRefreshNativeAuthTokenForNetwork } from './hooks';
 import { Dropdown, DropdownOption } from '../Dropdown';
 
 export const NetworkSwitcher = () => {
   const { activeNetwork } = useSelector(networkSelector);
-  const dispatch = useDispatch();
+  const refreshNativeAuthTokenForNetwork =
+    useRefreshNativeAuthTokenForNetwork();
+
   const networkOptions = networks.map((network) => ({
     label: network.name,
     value: network.id
@@ -17,7 +19,7 @@ export const NetworkSwitcher = () => {
     value: activeNetwork.id
   };
 
-  const handleNetworkSwitch = (option: DropdownOption) => {
+  const handleNetworkSwitch = async (option: DropdownOption) => {
     const selectedNetwork = networks.find(
       (network) => network.id === option.value
     );
@@ -26,7 +28,11 @@ export const NetworkSwitcher = () => {
       return;
     }
 
-    dispatch(changeNetwork(selectedNetwork));
+    await refreshNativeAuthTokenForNetwork({
+      networkId: selectedNetwork.id,
+      origin: window.location.origin,
+      signMessageCallback: (messageToSign) => Promise.resolve(messageToSign)
+    });
   };
 
   return (