+++ src/config/config.testnet.ts
@@ -10,18 +10,22 @@ export const networks: NetworkType[] = [
     name: 'Devnet',
     apiAddress: 'https://devnet-api.multiversx.com',
     gatewayUrl: '',
+    extrasApi: 'https://devnet-extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://devnet-api.multiversx.com'],
     sovereignContractAddress: '',
+    walletAddress: 'https://devnet-wallet.multiversx.com',
     WEGLDid: ''
   },
   {
-    default: true,
+    default: false,
     id: EnvironmentsEnum.devnet,
     name: 'Gateway',
     apiAddress: '',
     gatewayUrl: 'https://devnet-gateway.multiversx.com',
+    extrasApi: 'https://devnet-extras-api.multiversx.com',
     sampleAuthenticatedDomains: [''],
     sovereignContractAddress: '',
+    walletAddress: 'https://devnet-wallet.multiversx.com',
     WEGLDid: ''
   },
   {
@@ -30,8 +34,10 @@ export const networks: NetworkType[] = [
     name: 'Mainnet',
     apiAddress: 'https://api.multiversx.com',
     gatewayUrl: '',
+    extrasApi: 'https://extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://api.multiversx.com'],
     sovereignContractAddress: '',
+    walletAddress: 'https://wallet.multiversx.com',
     WEGLDid: ''
   },
   {
@@ -40,9 +46,11 @@ export const networks: NetworkType[] = [
     name: 'Sovereign',
     apiAddress: 'https://api-sovereign-test.elrond.ro',
     gatewayUrl: '',
+    extrasApi: 'https://extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://api-sovereign-test.elrond.ro'],
     sovereignContractAddress:
       'erd1qqqqqqqqqqqqqpgqfcm6l6rd42hwhskmk4thlp9kz58npfq50gfqdrthqa',
+    walletAddress: 'https://wallet.voyager1.dev',
     WEGLDid: 'WEGLD-bd4d79'
   },
   {
@@ -51,8 +59,10 @@ export const networks: NetworkType[] = [
     name: 'Testnet',
     apiAddress: 'https://testnet-api.multiversx.com',
     gatewayUrl: '',
+    extrasApi: 'https://testnet-extras-api.multiversx.com',
     sampleAuthenticatedDomains: ['https://testnet-api.multiversx.com'],
     sovereignContractAddress: '',
+    walletAddress: 'https://testnet-wallet.multiversx.com',
     WEGLDid: ''
   }
 ];