+++ src/helpers/api/index.ts
@@ -1,2 +1,5 @@
+export * from './getAxiosConfig';
 export * from './getBaseURL';
 export * from './getCurrentNetwork';
+export * from './getExtrasApi';
+export * from './retry';