+++ src/helpers/index.ts
@@ -1,4 +1,5 @@
 export * from './api';
 export * from './app';
+export * from './misc';
 export * from './tokens';
 export * from './validation';