+++ src/lib/sdkDapp.ts
@@ -63,3 +63,10 @@ export {
   setTransactionsToSignedState
 } from '@multiversx/sdk-dapp/services/transactions/updateSignedTransactions';
 export { sendBatchTransactions } from '@multiversx/sdk-dapp/services/transactions/sendBatchTransactions';
+export { useAxiosInterceptorContext } from '@multiversx/sdk-dapp/wrappers/AxiosInterceptorContext';
+export { storage } from '@multiversx/sdk-dapp/utils/storage';
+export { addNewCustomToast } from '@multiversx/sdk-dapp/utils/toasts';
+export {
+  maxDecimals,
+  stringIsFloat
+} from '@multiversx/sdk-dapp/utils/validation';